+++ provisioners/ansible/library/cf_security_groups.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python
 
 # Add security groups to CloudFormation template files.
-# This will modify the CloudFormation template files,
+# This will modify the CloudFormation template files.
 
 import sys, json, glob
 from ansible.module_utils.basic import *
@@ -20,13 +20,13 @@ def write_template(file_name, template):
         f.write(text)
 
 
-def add_security_groups_to_elb(elb_resource, security_groups, template):
+def add_security_groups_to_resource(resource_name, security_groups, template):
 
     for resource in template['Resources']:
-        if elb_resource in resource:
-            elb_security_groups = template['Resources'][elb_resource]['Properties']['SecurityGroups']
+        if resource_name in resource:
+            resource_security_groups = template['Resources'][resource_name]['Properties']['SecurityGroups']
             for secgrp in security_groups:
-                elb_security_groups.append(secgrp)
+                resource_security_groups.append(secgrp)
 
     return template
 
@@ -36,19 +36,19 @@ def main():
     module = AnsibleModule(
       argument_spec = dict(
         template_dir = dict(required=True, type='str'),
-        elb_resource = dict(required=True, type='str'),
+        resource_name = dict(required=True, type='str'),
         security_groups = dict(required=True, type='list')
       )
     )
 
     template_dir = module.params['template_dir']
-    elb_resource = module.params['elb_resource']
+    resource_name = module.params['resource_name']
     security_groups = module.params['security_groups']
 
     template_files = glob.glob(template_dir + "*.yaml")
     for template_file in template_files:
         template = read_template(template_file)
-        template = add_security_groups_to_elb(elb_resource, security_groups, template)
+        template = add_security_groups_to_resource(resource_name, security_groups, template)
         write_template(template_file, template)
 
     module.exit_json(changed = True, message = ", ".join(template_files))